Fix codacy check
