+++ HoaryFox/HoaryFox.csproj
@@ -60,6 +60,7 @@
     <Compile Include="Member\FrameBreps.cs" />
     <Compile Include="Member\CreateTag.cs" />
     <Compile Include="Member\FramePoints.cs" />
+    <Compile Include="Member\ShapeInfo.cs" />
     <Compile Include="Member\SteelCroSecBrep.cs" />
     <Compile Include="STBReader\STBReader\Member\Base.cs" />
     <Compile Include="STBReader\STBReader\Member\Enum.cs" />
